feat(cli): always start in INSERT mode when vim mode is enabled
The benefits of using Vim mode are limited to the input in the text
buffer, and is typically empty when the CLI is first started anyway.
Thus, there is no benefit in starting the CLI in NORMAL mode only
because vim mode is enabled.
On the contrary, it introduces friction in using the CLI with vim-mode
enabled, since users now need to explicitly enter INSERT mode first,
before they can provide input into the text buffer.

So, always start in INSERT mode when vim mode is enabled.